#1239 - Rename NYTCollectionReader to NitfReader
- Renamed module
- Renamed packages
- Removed NYT from type names
- Changed name of feature from "Guid" to "guid"
- Reformatting / cleaning up code
