+++ dkpro-core-asl/pom.xml
@@ -298,7 +298,7 @@
       </dependency>
       <dependency>
         <groupId>org.dkpro.core</groupId>
-        <artifactId>dkpro-core-io-nyt-asl</artifactId>
+        <artifactId>dkpro-core-io-nitf-asl</artifactId>
         <version>1.11.0-SNAPSHOT</version>
       </dependency>
       <dependency>
@@ -572,7 +572,7 @@
     <module>../dkpro-core-io-lxf-asl</module>
     <module>../dkpro-core-io-negra-asl</module>
     <module>../dkpro-core-io-nif-asl</module>
-    <module>../dkpro-core-io-nyt-asl</module>
+    <module>../dkpro-core-io-nitf-asl</module>
     <module>../dkpro-core-io-pdf-asl</module>
     <module>../dkpro-core-io-penntree-asl</module>
     <module>../dkpro-core-io-pubannotation-asl</module>
