+++ dkpro-core-io-nitf-asl/LICENSE.txt
