+++ dkpro-core-io-nitf-asl/pom.xml
@@ -25,7 +25,7 @@
     <version>1.11.0-SNAPSHOT</version>
     <relativePath>../dkpro-core-asl</relativePath>
   </parent>
-  <artifactId>dkpro-core-io-nyt-asl</artifactId>
+  <artifactId>dkpro-core-io-nitf-asl</artifactId>
   <packaging>jar</packaging>
   <name>DKPro Core ASL - IO - New York Times Corpus</name>
   <url>https://dkpro.github.io/dkpro-core/</url>
