+++ dkpro-core-io-nitf-asl/src/filter/filter.properties
