+++ dkpro-core-io-nitf-asl/src/main/resources/META-INF/org.apache.uima.fit/types.txt
@@ -1,0 +1,1 @@
+classpath*:desc/type/ArticleMetaData.xml
