+++ dkpro-core-io-nitf-asl/src/test/resources/data/1987/01/01/0000000.xml
