+++ dkpro-core-io-nitf-asl/src/test/resources/data/1987/02/01/0000002.xml
