+++ dkpro-core-io-nitf-asl/src/test/resources/data/1988/01/01/0000003.xml
