--- dkpro-core-io-nyt-asl/src/main/resources/META-INF/org.apache.uima.fit/types.txt
@@ -1,1 +1,0 @@
-classpath*:desc/type/NYTArticleMetaData.xml
