--- dkpro-core-io-nyt-asl/src/main/resources/desc/type/NYTArticleMetaData.xml
@@ -1,211 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<typeSystemDescription xmlns="http://uima.apache.org/resourceSpecifier">
-	<name>NYTArticleMetaData</name>
-	<description/>
-  <version>${version}</version>
-  <vendor>Ubiquitous Knowledge Processing (UKP) Lab, Technische Universität Darmstadt</vendor>
-	<types>
-		<typeDescription>
-			<name>org.dkpro.core.io.nyt.metadata.NYTArticleMetaData</name>
-			<description>A document annotation that describes the metadata of a
-				newspaper article.</description>
-			<supertypeName>uima.cas.AnnotationBase</supertypeName>
-			<features>
-				<featureDescription>
-					<name>Guid</name>
-					<description>The GUID field specifies a (4-byte) integer that is
-						guaranteed
-						to be unique for every document
-						in the corpus.</description>
-					<rangeTypeName>uima.cas.Integer</rangeTypeName>
-				</featureDescription>
-				<featureDescription>
-					<name>alternateUrl</name>
-					<description>This field specifies the location on nytimes.com of
-						the article. When present, this URL is preferred to the URL field
-						on articles published on or after April 02,
-						2006, as the linked
-						page will have richer content.</description>
-					<rangeTypeName>uima.cas.String</rangeTypeName>
-				</featureDescription>
-				<featureDescription>
-					<name>url</name>
-					<description>This field specifies the location on nytimes.com of
-						the article. The 'Alternative Url'
-						field is preferred to this field
-						on articles published on or after
-						April 02, 2006, as the
-						linked page
-						will have richer content.</description>
-					<rangeTypeName>uima.cas.String</rangeTypeName>
-				</featureDescription>
-				<featureDescription>
-					<name>publicationDate</name>
-					<description>This field specifies the date of the article's
-						publication. This field is specified in the
-						format
-						YYYYMMDD'T'HHMMSS where:
-						1. YYYY is the four-digit year.
-						2. MM is
-						the two-digit month [01-12].
-						3. DD is the two-digit day [01-31].
-						4.
-						T is a constant value.
-						5. HH is the two-digit hour [00-23].
-						6. MM is
-						the two-digit minute-past-the hour [00-59]
-						7. SS is the two-digit
-						seconds-past-the-minute [00-59].
-						Please note that values for HH,MM,
-						and SS are not defined for this
-						corpus, that is to day
-						HH,MM, and SS
-						are always defined to be '00'.</description>
-					<rangeTypeName>uima.cas.String</rangeTypeName>
-				</featureDescription>
-				<featureDescription>
-					<name>typesOfMaterial</name>
-					<description>This field specifies a normalized list of terms
-						describing the general editorial category of the article.
-						These
-						tags are algorithmically assigned and
-						manually verified by
-						nytimes.com production staff.
-						Examples Include:
-						* REVIEW
-						* OBITUARY
-						* ANALYSIS</description>
-					<rangeTypeName>uima.cas.StringArray</rangeTypeName>
-				</featureDescription>
-				<featureDescription>
-					<name>headline</name>
-					<description>This field specifies the headline of the article as it
-						appeared in the
-						print edition of the New York
-						Times.</description>
-					<rangeTypeName>uima.cas.String</rangeTypeName>
-				</featureDescription>
-				<featureDescription>
-					<name>onlineHeadline</name>
-					<description>This field specifies the headline displayed with the
-						article on
-						nytimes.com. Often
-						this differs from the headline used in
-						print.</description>
-					<rangeTypeName>uima.cas.String</rangeTypeName>
-				</featureDescription>
-				<featureDescription>
-					<name>columnName</name>
-					<description>If the article is part of a regular column, this field
-						specifies the
-						name of that column.
-						Sample Column Names:
-						1. World News
-						Briefs
-						2. WEDDINGS
-						3. The Accessories Channel</description>
-					<rangeTypeName>uima.cas.String</rangeTypeName>
-				</featureDescription>
-				<featureDescription>
-					<name>author</name>
-					<description>This field is based on the normalized byline in the
-						original corpus data: "The Normalized Byline field is the byline
-						normalized to the form (last name, first
-						name)".</description>
-					<rangeTypeName>uima.cas.String</rangeTypeName>
-				</featureDescription>
-				<featureDescription>
-					<name>descriptors</name>
-					<description>The 'descriptors' field specifies a list of
-						descriptive terms drawn
-						from a normalized controlled
-						vocabulary
-						corresponding to subjects mentioned in the article. These tags
-						are
-						hand-assigned by
-						a team of library scientists working in the New
-						York Times Indexing
-						service.
-						Examples Include:
-						* ECONOMIC CONDITIONS
-						AND TRENDS
-						* AIRPLANES
-						* VIOLINS</description>
-					<rangeTypeName>uima.cas.StringArray</rangeTypeName>
-				</featureDescription>
-				<featureDescription>
-					<name>onlineDescriptors</name>
-					<description>This field specifies a list of descriptors from a
-						normalized
-						controlled
-						vocabulary that
-						correspond to topics mentioned
-						in the article. These
-						tags are
-						algorithmically
-						assigned and manually
-						verified by
-						nytimes.com production staff.
-						Examples Include:
-						* Marriages
-						* Parks and Other Recreation Areas
-						* Cooking and Cookbooks</description>
-					<rangeTypeName>uima.cas.StringArray</rangeTypeName>
-				</featureDescription>
-				<featureDescription>
-					<name>generalOnlineDescriptors</name>
-					<description>The 'general online descriptors' field specifies a
-						list of descriptors that are at a higher level of
-						generality than
-						the other tags associated with the article. These tags are
-						algorithmically
-						assigned and manually verified by nytimes.com
-						production staff.
-						Examples Include:
-						* Surfing
-						* Venice Biennale
-						* Ranches</description>
-					<rangeTypeName>uima.cas.String</rangeTypeName>
-				</featureDescription>
-				<featureDescription>
-					<name>onlineSection</name>
-					<description>This field specifies the section(s) on nytimes.com in
-						which the
-						article is placed. If
-						the article is placed in multiple
-						sections, this field will be
-						specified as a ';' delineated
-						list.</description>
-					<rangeTypeName>uima.cas.String</rangeTypeName>
-				</featureDescription>
-				<featureDescription>
-					<name>section</name>
-					<description>This field specifies the section of the paper in which
-						the article
-						appears. This is not
-						the name of the section, but rather
-						a letter or number that indicates
-						the section.</description>
-					<rangeTypeName>uima.cas.String</rangeTypeName>
-				</featureDescription>
-				<featureDescription>
-					<name>taxonomicClassifiers</name>
-					<description>This field specifies a list of taxonomic classifiers
-						that place this
-						article into a
-						hierarchy of articles. The individual
-						terms of each taxonomic classifier
-						are separated with the '/' character. 
-						These tags are algorithmically assigned and manually
-						verified
-						by nytimes.com production staff.
-						Examples Include:
-						* Top/Features/Travel/Guides/Destinations/North America/United States/Arizona
-						* Top/News/U.S./Rockies
-						* Top/Opinion</description>
-					<rangeTypeName>uima.cas.StringArray</rangeTypeName>
-				</featureDescription>
-			</features>
-		</typeDescription>
-	</types>
-</typeSystemDescription>
